// projects/ngx-sp-auth/src/lib/components/menu-lateral/model/ret-infrausuarioimg.ts
import { IError } from "ngx-sp-infra";
import { InfraUsuarioImg } from './infrausuarioimg';

export class RetInfraUsuarioImg implements IError{
  Error: boolean = false;
  ErrorMessage: string = "";
  InfraUsuarioImg: InfraUsuarioImg = new InfraUsuarioImg();
  uncompressedFileSize?: number | null = 0;
  compressedFileSize?: number | null = 0;
}